requirements.txt
blinker==1.6.3
build==1.0.3
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
elastic-transport==8.4.1
elasticsearch==8.10.1
filelock==3.13.1
Flask==3.0.0
fsspec==2023.10.0
huggingface-hub==0.17.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.1
packaging==23.2
Pillow==10.1.0
pip-tools==7.3.0
pyproject_hooks==1.0.0
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.14.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.35.0
typing_extensions==4.8.0
urllib3==1.26.18
Werkzeug==3.0.1
